Back to Unity 2021
Unity 2022 has recently deprecated certain methods used in Kit.AR. As a result, we have reverted to the 2021 version and will return to 2022 once we are ready. We have made some enhancements to the video montage feature, which now automatically exits once the video has ended. In the AR scene, the toggle button for tracking markers has been removed. Tracking markers will now be automatically dismissed once an object is placed on the ground.